refactor: :recycle: refactoring attributes and folder hierarchy
