+++ src/main/java/com/u2team/huansync/activity/cosplay/categorycosplay/model/CategoryCosplayDAO.java
@@ -1,4 +1,4 @@
-package com.u2team.huansync.activity.cosplay.categorycosplay.controller;
+package com.u2team.huansync.activity.cosplay.categorycosplay.model;
 
 import com.u2team.huansync.activity.cosplay.categorycosplay.model.CategoryCosplay;
 import com.u2team.huansync.persistence.BDConnection;
